backend/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
python-multipart==0.0.20

# Database
supabase==2.9.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.26.0  # For testing FastAPI endpoints
websockets==12.0  # For testing WebSocket endpoints

# Additional utilities
pydantic==2.12.3
pydantic-settings==2.1.0

# AI Services
openai>=1.0.0
google-generativeai>=0.3.0

# Celery & Redis (Task Queue for MCP Agent)
celery[redis]==5.3.4
redis>=4.5.2,<5.0.0  # Compatible with celery[redis]
flower==2.0.1  # Optional monitoring UI

# Entity Extraction (Story 8.2)
phonenumbers==8.13.27  # Phone number parsing and validation

# Domain Reputation Tool (Story 8.5)
python-whois==0.9.4  # WHOIS lookup for domain age

